+++ lib/image_processing/mini_magick.rb
@@ -57,12 +57,15 @@ module ImageProcessing
         magick.merge! args
       end
 
-      def load_image(path_or_magick, page: nil, geometry: nil, fail: true, auto_orient: true)
+      def load_image(path_or_magick, page: nil, geometry: nil, fail: true, auto_orient: true, define: {}, **options)
         if path_or_magick.is_a?(::MiniMagick::Tool)
           magick = path_or_magick
         else
           source_path = path_or_magick
           magick = ::MiniMagick::Tool::Convert.new
+
+          apply_define(magick, define)
+          apply_options(magick, options)
 
           input_path  = source_path
           input_path += "[#{page}]" if page
@@ -77,9 +80,37 @@ module ImageProcessing
         magick
       end
 
-      def save_image(magick, destination_path, **options)
+      def save_image(magick, destination_path, define: {}, **options)
+        apply_define(magick, define)
+        apply_options(magick, options)
+
         magick << destination_path
+
         magick.call
+      end
+
+      private
+
+      def apply_define(magick, define)
+        define.each do |namespace, options|
+          namespace = namespace.to_s.gsub("_", "-")
+
+          options.each do |key, value|
+            key = key.to_s.gsub("_", "-")
+
+            magick.define "#{namespace}:#{key}=#{value}"
+          end
+        end
+      end
+
+      def apply_options(magick, options)
+        options.each do |option, value|
+          case value
+          when true  then magick.send(option)
+          when false then magick.send(option).+
+          else            magick.send(option, *value)
+          end
+        end
       end
     end